Added new icon - PGP-key
New icon has been added for 'PGP-key'.

It may be linked to a keystore such as: https://keys.openpgp.org

...

This PR also contains a couple of minor fixes to the ReadMe...

'Pinterest' wasn't displayed due to it being in the 5th column of the 4 column table.

Removed unnecessary word 'be'.